requirements.txt
python-dotenv
PyQt5
kiteconnect
